refactor(solver): extract smt_solver.py, smt.py becomes re-export layer
Move SMTSolver class and checkpoint function from smt.py into a new
smt_solver.py module. Replace smt.py with a thin re-export layer that
imports from smt_types, smt_translate, and smt_solver for backward
compatibility. Update smt_translate.py TYPE_CHECKING import to point
to the new location.
